requirements.txt
# Core dependencies
pip>=25.0.1
setuptools>=69.0.3
wheel>=0.42.0

# Trading and Data Analysis
MetaTrader5>=5.0.45  # MT5 API
pandas>=2.2.0        # Data manipulation
numpy>=1.26.0       # Numerical computations
ta-lib>=0.4.28      # Technical analysis
backtrader>=1.9.78  # Backtesting framework

# Database
psycopg2>=2.9.9     # PostgreSQL adapter
SQLAlchemy>=2.0.27  # ORM
alembic>=1.13.1     # Database migrations

# Web Framework and Dashboard
Flask>=3.0.2        # Web framework
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3  # User authentication
Flask-WTF>=1.2.1    # Forms
dash>=2.14.2        # Interactive dashboards
plotly>=5.18.0      # Plotting

# API and Integration
requests>=2.31.0    # HTTP requests
python-telegram-bot>=21.11.1  # Telegram integration
tweepy>=4.14.0      # Twitter API
websocket-client>=1.7.0  # WebSocket support

# Machine Learning (Optional)
scikit-learn>=1.4.0
tensorflow>=2.15.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development and Utilities
python-dotenv>=1.0.1  # Environment variables
black>=24.1.1        # Code formatting
flake8>=7.0.0       # Linting
mypy>=1.8.0         # Type checking
jupyter>=1.0.0      # Notebooks for analysis

# Documentation
Sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0

# Monitoring and Logging
grafana>=9.5.0
prometheus-client>=0.19.0

# Added seaborn
seaborn>=0.13.0 